% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{eigen_table}
\alias{eigen_table}
\title{Generalized eigenvalues of a solved model}
\usage{
eigen_table(x)
}
\arguments{
\item{x}{A \code{qpm_solution}.}
}
\value{
A data frame with the moduli of the generalized eigenvalues of
the model companion pencil, sorted ascending, with stability flags.
}
\description{
Generalized eigenvalues of a solved model
}
\examples{
head(eigen_table(qpm_solve(qpm_template("bkl"))))
}
