% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs.qpm_filtration}
\alias{nobs.qpm_filtration}
\alias{nobs.qpm_estimate}
\title{Number of observations}
\usage{
\method{nobs}{qpm_filtration}(object, ...)

\method{nobs}{qpm_estimate}(object, ...)
}
\arguments{
\item{object}{A \code{qpm_filtration} or \code{qpm_estimate}.}

\item{...}{Unused.}
}
\value{
Number of time periods used.
}
\description{
Number of observations
}
\examples{
sol <- qpm_solve(qpm_template("bkl"))
obs <- simulate(sol, nsim = 40, seed = 1, burn = 20)
nobs(qpm_filter(sol, obs[, c("period", "pi", "i")]))
}
