% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{steady_state}
\alias{steady_state}
\title{Steady state of a model or solution}
\usage{
steady_state(x, ...)
}
\arguments{
\item{x}{A \code{qpm_model} or \code{qpm_solution}.}

\item{...}{Unused.}
}
\value{
Named vector of steady-state values for the declared variables.
}
\description{
Steady state of a model or solution
}
\examples{
steady_state(qpm_template("bkl"))
}
